Just before going to check if the current bugg is due to that there are no more bumps. So we are going to add a new bump to the initial configuration of the road

+++ Discrete approach/DiscreteSimulation.h
@@ -7,7 +7,7 @@ typedef struct {
 
 typedef struct {
 	unsigned short height; // How many blocks has the pile 'f_max'
-	unsigned short f; //
+	int f; //
 	BLOCK * blocks; //BLOCKs of the pile, 
 } PILE;
